backend/requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
PyGithub>=2.8.1
openai>=2.7.1
python-dotenv>=1.2.1
requests>=2.32.5
gunicorn>=23.0.0
structlog>=24.1.0
httpx<0.28
docker>=7.1.0
boto3>=1.34.0
groq>=0.9.0
supabase>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
redis>=5.0.1
rq>=1.16.2
